backend/requirements.txt
fastapi
uvicorn
sqlalchemy
pydantic
pydantic-settings
asyncpg
python-dotenv
httpx
pyjwt
